Renamed members of structure band

+++ descriptor/include/Hist.hpp
@@ -14,7 +14,7 @@ enum Dimension
 
 struct Bins
 {
-	std::vector<double> bins;
+	std::vector<float> bins;
 	double step;
 	Dimension dimension;
